+++ platformio/fs.py
@@ -24,7 +24,7 @@ import sys
 
 import click
 
-from platformio import exception
+from platformio import exception, proc
 from platformio.compat import IS_WINDOWS
 
 
@@ -41,7 +41,7 @@ class cd(object):
 
 
 def get_source_dir():
-    curpath = os.path.realpath(__file__)
+    curpath = os.path.abspath(__file__)
     if not os.path.isfile(curpath):
         for p in sys.path:
             if os.path.isfile(os.path.join(p, __file__)):
@@ -119,7 +119,7 @@ def ensure_udev_rules():
     if not any(os.path.isfile(p) for p in installed_rules):
         raise exception.MissedUdevRules
 
-    origin_path = os.path.realpath(
+    origin_path = os.path.abspath(
         os.path.join(get_source_dir(), "..", "scripts", "99-platformio-udev.rules")
     )
     if not os.path.isfile(origin_path):
@@ -179,6 +179,25 @@ def to_unix_path(path):
     if not IS_WINDOWS or not path:
         return path
     return re.sub(r"[\\]+", "/", path)
+
+
+def normalize_path(path):
+    path = os.path.abspath(path)
+    if not IS_WINDOWS or not path.startswith("\\\\"):
+        return path
+    try:
+        result = proc.exec_command(["net", "use"])
+        if result["returncode"] != 0:
+            return path
+        share_re = re.compile(r"\s([A-Z]\:)\s+(\\\\[^\s]+)")
+        for line in result["out"].split("\n"):
+            share = share_re.search(line)
+            if not share:
+                continue
+            path = path.replace(share.group(2), share.group(1))
+    except OSError:
+        pass
+    return path
 
 
 def expanduser(path):